Django-Middleware-0x03/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==3.2.25
django-filter==23.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.93.4
pycodestyle==2.10.0
PyJWT==2.8.0
pytz==2025.2
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2025.2
